RavenDB-17601 Rebasing to null support
