+++ src/Corax/IndexEntryBuilder/IndexEntryFieldType.cs
@@ -1,0 +1,21 @@
+﻿using System;
+
+namespace Corax;
+
+[Flags]
+public enum IndexEntryFieldType : ushort
+{
+    Null = 0,
+    Simple = 1,
+    Tuple = 1 << 1,
+    List = 1 << 2,
+    Raw = 1 << 3,
+    SpatialPoint = 1 << 5,
+    EmptyList = 1 << 13,
+    HasNulls = 1 << 14, // Helper for list writer.
+    Invalid = 1 << 15,
+
+    SpatialPointList = List | SpatialPoint,
+    TupleList = List | Tuple,
+    RawList = List | Raw
+}
